restrict all types to clonable, to support downstream cloning
