+++ examples/custom_strategy.rs
@@ -1,10 +1,11 @@
 use ndarray::prelude::*;
 
+use ninterp::data::InterpData2D;
 use ninterp::prelude::*;
 use ninterp::strategy::*;
 
 // Debug must be derived for custom strategies
-#[derive(Debug)]
+#[derive(Debug, Clone)]
 struct CustomStrategy;
 
 // Implement strategy for 2-D f32 interpolation
@@ -14,7 +15,7 @@ where
     // e.g. `Array2<f32>`, `ArrayView2<f32>`, `CowArray<<'a, f32>, Ix2>`, etc.
     // For a more generic bound, consider introducing a bound for D::Elem
     // e.g. D::Elem: num_traits::Num + PartialOrd
-    D: ndarray::Data<Elem = f32>,
+    D: ndarray::Data<Elem = f32> + ndarray::RawDataClone,
 {
     fn interpolate(
         &self,
