+++ examples/uom.rs
@@ -12,7 +12,7 @@ fn main() {
     let f_x = array![Power::new::<kilowatt>(0.25), Power::new::<kilowatt>(0.75)];
     // `uom::si::Quantity` is repr(transparent), meaning it has the same memory layout as its contained type.
     // This means we can get the contained type via transmuting.
-    let interp: Interp1D<ndarray::OwnedRepr<f64>, _> = unsafe {
+    let interp: ninterp::one::Interp1DOwned<f64, _> = unsafe {
         Interp1D::new(
             std::mem::transmute(x),
             std::mem::transmute(f_x),
