+++ src/lib.rs
@@ -123,6 +123,7 @@ pub mod prelude {
     pub use crate::Interpolator;
 }
 
+pub mod data;
 pub mod error;
 pub mod strategy;
 
@@ -133,20 +134,22 @@ pub mod two;
 pub mod zero;
 
 pub mod interpolator {
-    pub use crate::n::InterpND;
-    pub use crate::one::Interp1D;
-    pub use crate::three::Interp3D;
-    pub use crate::two::Interp2D;
+    pub use crate::n::{InterpND, InterpNDOwned, InterpNDViewed};
+    pub use crate::one::{Interp1D, Interp1DOwned, Interp1DViewed};
+    pub use crate::three::{Interp3D, Interp3DOwned, Interp3DViewed};
+    pub use crate::two::{Interp2D, Interp2DOwned, Interp2DViewed};
     pub use crate::zero::Interp0D;
 }
 
+pub(crate) use data::*;
 pub(crate) use error::*;
 pub(crate) use strategy::*;
 
 pub(crate) use std::fmt::Debug;
 
+pub use ndarray;
 pub(crate) use ndarray::prelude::*;
-pub(crate) use ndarray::{Data, Ix};
+pub(crate) use ndarray::{Data, Ix, RawDataClone};
 
 pub(crate) use num_traits::{clamp, Num, One};
 
@@ -194,54 +197,6 @@ impl<T> Interpolator<T> for Box<dyn Interpolator<T>> {
     }
 }
 
-#[derive(Debug)]
-#[cfg_attr(feature = "serde", derive(Deserialize, Serialize))]
-#[cfg_attr(
-    feature = "serde",
-    serde(bound = "
-        D: DataOwned,
-        D::Elem: Serialize + DeserializeOwned,
-        Dim<[usize; N]>: Serialize + DeserializeOwned,
-        [ArrayBase<D, Ix1>; N]: Serialize + DeserializeOwned,
-    ")
-)]
-pub struct InterpData<D, const N: usize>
-where
-    Dim<[Ix; N]>: Dimension,
-    D: Data,
-    D::Elem: Num + PartialOrd + Copy + Debug,
-{
-    pub grid: [ArrayBase<D, Ix1>; N],
-    pub values: ArrayBase<D, Dim<[Ix; N]>>,
-}
-
-impl<D, const N: usize> InterpData<D, N>
-where
-    Dim<[Ix; N]>: Dimension,
-    D: Data,
-    D::Elem: Num + PartialOrd + Copy + Debug,
-{
-    pub fn validate(&self) -> Result<(), ValidateError> {
-        for i in 0..N {
-            let i_grid_len = self.grid[i].len();
-            // Check that each grid dimension has elements
-            // Indexing `grid` directly is okay because empty dimensions are caught at compilation
-            if i_grid_len == 0 {
-                return Err(ValidateError::EmptyGrid(i));
-            }
-            // Check that grid points are monotonically increasing
-            if !self.grid[i].windows(2).into_iter().all(|w| w[0] <= w[1]) {
-                return Err(ValidateError::Monotonicity(i));
-            }
-            // Check that grid and values are compatible shapes
-            if i_grid_len != self.values.shape()[i] {
-                return Err(ValidateError::IncompatibleShapes(i));
-            }
-        }
-        Ok(())
-    }
-}
-
 /// Extrapolation strategy
 ///
 /// Controls what happens if supplied interpolant point
@@ -264,7 +219,7 @@ macro_rules! extrapolate_impl {
     ($InterpType:ident, $Strategy:ident) => {
         impl<D, S> $InterpType<D, S>
         where
-            D: Data,
+            D: Data + RawDataClone,
             D::Elem: Num + PartialOrd + Copy + Debug,
             S: $Strategy<D>,
         {
