+++ src/two/strategies.rs
@@ -2,7 +2,7 @@ use super::*;
 
 impl<D> Strategy2D<D> for Linear
 where
-    D: Data,
+    D: Data + RawDataClone,
     D::Elem: Num + PartialOrd + Copy + Debug,
 {
     fn interpolate(
@@ -51,7 +51,7 @@ where
 
 impl<D> Strategy2D<D> for Nearest
 where
-    D: Data,
+    D: Data + RawDataClone,
     D::Elem: Num + PartialOrd + Copy + Debug,
 {
     fn interpolate(
